Improve typing for authentication related resources
